refac : note, user 서비스 파일 수정
