+++ src/api/noteService.js
@@ -1,5 +1,17 @@
 import API from "../axios";
 
+// 노트 조회하기 (GET)
+export const getNotes = async (userID) => {
+    const resp = await API.get("/note/preview");
+    return resp;
+};
+
+// 노트 상세 조회하기 (GET)
+export const getNote = async (noteId) => {
+    const resp = await API.get(`/note/detail?noteId=${noteId}`);
+    return resp;
+};
+
 // 노트 저장하기 (POST)
 export const saveNote = async (body) => {
     return API.post("/note", body, {
