+++ src/api/userService.js
@@ -13,18 +13,6 @@ export const checkToken = async (token) => {
     return resp;
 };
 
-// 노트 조회하기 (GET)
-export const getNotes = async (userID) => {
-    const resp = await API.get("/note/preview");
-    return resp;
-};
-
-// 노트 상세 조회하기 (GET)
-export const getNote = async (noteId) => {
-    const resp = await API.get(`/note/detail?noteId=${noteId}`);
-    return resp;
-};
-
 // 연락 폼 보내기 (POST)
 export const sendContact = async (name, email, message) => {
     const req = {
